Remove subscription stuff
Subscription doesn't seem to be a viable thing for this, no one is using
it. Makes sense, it's a safari extension and doesn't have a ton of
week-by-week value. I'll just make it one time fee.

Removes native store view, removes all the activation stuff and
verifying an active subscription.